refactor(insights): compute lap insights server-side
Extract pure physics to shared/lib/vehicle-physics.ts, move lap-insights to shared/lib, serve insights in GET /api/laps/:id, AI prompts import shared.
